+++ delphin/tdl/_model.py
@@ -1,6 +1,6 @@
 from collections.abc import Mapping, Sequence
 from pathlib import Path
-from typing import TypeAlias, Union
+from typing import NamedTuple, TypeAlias, Union
 
 from delphin import util
 from delphin.tdl._exceptions import TDLError
@@ -865,6 +865,36 @@ class InstanceEnvironment(_Environment):
     def __init__(self, status, entries=None):
         super().__init__(entries)
         self.status = status
+
+
+class ConfigEnvironment(_Environment):
+    """
+    TDL configuration environment.
+
+    Args:
+        entries (list): config entries
+    """
+
+    def __init__(self, label: str = "", entries=None):
+        super().__init__(entries)
+        self.label = label
+
+
+class ConfigEntry(NamedTuple):
+    """Key-value pair from a TDL config file.
+
+    Since the type of the value (e.g., an atomic string or a list)
+    depends on the key, values are always read as a list. For
+    convenience, the :prop:`value` property returns these values
+    as a single string joined with space characters.
+    """
+
+    key: str
+    values: list[str]
+
+    @property
+    def value(self) -> str:
+        return " ".join(self.values)
 
 
 class FileInclude: